refactor(matrix): extract matrix utilities into subpackage and facade
- introduce mmrelay.matrix package with __init__.py, prefixes.py and
  room_mapping.py
- refactor matrix_utils to delegate core logic to room_mapping via
  lazy imports
- establish facade exports and prepare explicit submodule exports
