+++ src/mmrelay/matrix/__init__.py
@@ -1,0 +1,5 @@
+"""Matrix utility submodules - facade re-exports."""
+
+# This package is the home for decomposed Matrix utilities.
+# The matrix_utils.py module remains the state owner and facade.
+# Submodules access shared state via: import mmrelay.matrix_utils as facade
